Set a more sensible repl-mode by default
Only set `tty` as the default when actually running under a TTY.

This fixes fallout effects such as the issues in Jupyter::Chatbook
where starting a REPL from inside a script could not be effected
without the addition of a new 'managed' repl-mode.

This seems simpler both implementation-wise and conceptually, but
the 'managed' repl-mode can be revisited at a later time should
this patch not end up covering all corner cases.