Fix missing register initialization in repeat loops
This is finally a proper fix replacing the workaround introduced in
commit 2f2ac2f0665980ef54e2bd8a634ad5ff0a590b4a.

The problem was that for loops for which we need to pass the condition
to the block (think while $cond -> $foo { }) we only compile the
condition handling QAST to MAST very late, so at the time when we need
to emit the nulling of the temporary condition register, we didn't have
that register allocated yet. The workaround was just to not null the
register and hope for the best. This worked for more than 6 years (at
least as far as we know) but it finally surfaced now. Time for a proper
fix.

Fix by manually allocating the register instead of relying on the
QASTCompilerMAST to do it for a variable declaration. This way we can
control the timing better.