Ensure we declare our dynamics even when there's no compunit
Naughty users (like Rakudo's ProxyReaderFactory dispatch helper) might
compile a block without enclosing it in a compunit. In that case we
wouldn't declare the %*BLOCKS_DONE and $*BLOCK dynamic variables which
may lead to them getting auto-vivified in the package (and exported) and
may pick up confusing values from outside. Better declare them in a
place that's harder to avoid.